refactor: reorganize handler logic and move to dedicated handlers file
